requirements.txt
requests
beautifulsoup4
pandas
pyyaml
typer
streamlit
pytest
